Separate functions into separate files and update exports/imports
