+++ src/card.js
@@ -7,22 +7,6 @@ const createCard = (id, question, answers, correctAnswer) => {
   }
 }
 
-const evaluateGuess = (guess, correctAnswer) => {
- if(guess === correctAnswer){
-  return true;
- } else {
-  return false;
- }
-}
-
-const createDeck = (cards) => {
- return cards
-}
-
-const countCards = (deck) => {
-  return deck.length;
-}
-
 const createRound = (deck) => {
   return {
     deck: deck,
@@ -34,8 +18,5 @@ const createRound = (deck) => {
 
 module.exports = {
   createCard,
-  evaluateGuess,
-  createDeck,
-  countCards,
   createRound
 }
