+++ src/deck.js
@@ -1,0 +1,13 @@
+
+const createDeck = (cards) => {
+  return cards
+ }
+ 
+ const countCards = (deck) => {
+   return deck.length;
+ }
+
+ module.exports = {
+  createDeck,
+  countCards
+ }
