+++ src/turns.js
@@ -1,0 +1,11 @@
+const evaluateGuess = (guess, correctAnswer) => {
+  if(guess === correctAnswer){
+   return true;
+  } else {
+   return false;
+  }
+ }
+
+ module.exports = {
+  evaluateGuess
+ }
